Added CHAPTER #. to PDF bookmarks
I used \texorpdfstring in the title of each chapter to specify a different string to be used in the PDF bookmarks.  This string references the chapter counter to get the correct chapter number.

Long-term: might want to modify the Chapter code incorporate this so that the user doesn't have to.  Also need to modify Chapter so that extra space is not added between entries in the List of Figures/Tables when the chapter changes.